fix: address lazy bundle review feedback
